// frontend/src/models/car.ts
export type Car = {
    vin: string;
    image: string;
    manufacturer: string;
    model: string;
    constructionYear: number;
    mileage: number;
    engineSize: number;
    power: number;
    gearbox: string;
    fuelType: string;
    price: number;
    description: string;
    equipment: string;
};